Make classes 'static
This is basically to allow Instances to have a reference to a class
that outlives the lifetime of the instance itself.
I also removed the Callable trait and instead call into the struct
directly. This is mainly to only have the the lifetime on self for
the call method on Class.